klighd: use mock text bounds for LS.
